requirements.txt
# =============================
# requirements.txt
# =============================
# fastapi web framework
# uvicorn ASGI server (\"[standard]\" adds useful extras for dev)
# openai SDK v1
# python-dotenv for .env support
# httpx used by openai client
# pydantic comes with FastAPI, but listed for clarity
#
# Install with: pip install -r requirements.txt

fastapi>=0.110
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
python-multipart>=0.0.9
openai>=1.42
slowapi>=0.1.9
limits>=3.13
pdfplumber>=0.11

# Optional: for distributed rate limiting storage (instead of in-memory)
# redis==5.0.8# Optional: for distributed rate limiting storage (instead of in-memory)

# Data science stack
numpy
pandas
matplotlib

# Dev tools
black            # code formatter
ruff             # linter / static analysis

